corrigido o tipo do atributo preço.

+++ src/main/java/com/generation/lojagames/model/Produto.java
@@ -55,11 +55,11 @@ public class Produto {
         this.plataforma = plataforma;
     }
 
-    public Double getPreco() {
+    public BigDecimal getPreco() {
         return preco;
     }
 
-    public void setPreco(Double preco) {
+    public void setPreco(BigDecimal preco) {
         this.preco = preco;
     }
